+++ Runtime/ContinuousDistributions/ExponentialDistribution/GeneratorProviders/ExponentialGeneratorDependentSimpleProvider.cs
@@ -1,0 +1,134 @@
+// Copyright (c) 2020 Vladimir Popov zor1994@gmail.com https://github.com/ZorPastaman/Random-Generators
+
+using System;
+using System.Runtime.CompilerServices;
+using JetBrains.Annotations;
+using UnityEngine;
+
+namespace Zor.RandomGenerators.ContinuousDistributions
+{
+	/// <summary>
+	/// Provides <see cref="ExponentialGeneratorDependentSimple{T}"/>.
+	/// </summary>
+	[CreateAssetMenu(
+		menuName = CreateAssetMenuConstants.ExponentialDistributionFolder +
+			"Exponential Generator Dependent Simple Provider",
+		fileName = "Exponential Generator Dependent Simple Provider",
+		order = CreateAssetMenuConstants.DistributionOrder
+	)]
+	public sealed class ExponentialGeneratorDependentSimpleProvider : ContinuousGeneratorProvider
+	{
+#pragma warning disable CS0649
+		[SerializeField, Tooltip("Random generator that returns an independent and identically distributed random value in range [0, 1].")]
+		private ContinuousGeneratorProviderReference m_DependedGeneratorProvider;
+		[SerializeField, Tooltip("Mustn't be zero.")]
+		private float m_Lambda = ExponentialDistribution.DefaultLambda;
+#pragma warning restore CS0649
+
+		[NonSerialized] private ExponentialGeneratorDependentSimple<IContinuousGenerator> m_sharedGenerator;
+
+		/// <summary>
+		/// Creates a new <see cref="ExponentialGeneratorDependentSimple{T}"/>
+		/// and returns it as <see cref="IContinuousGenerator"/>.
+		/// </summary>
+		public override IContinuousGenerator generator
+		{
+			[Pure]
+			get => new ExponentialGeneratorDependentSimple<IContinuousGenerator>(
+				m_DependedGeneratorProvider.generator, m_Lambda);
+		}
+
+		/// <summary>
+		/// Returns a shared <see cref="ExponentialGeneratorDependentSimple{T}"/> as <see cref="IContinuousGenerator"/>.
+		/// </summary>
+		public override IContinuousGenerator sharedGenerator
+		{
+			get
+			{
+				if (m_sharedGenerator == null)
+				{
+					m_sharedGenerator = exponentialGenerator;
+				}
+
+				return m_sharedGenerator;
+			}
+		}
+
+		/// <summary>
+		/// Creates a new <see cref="ExponentialGeneratorDependentSimple{T}"/> and returns it.
+		/// </summary>
+		[NotNull]
+		public ExponentialGeneratorDependentSimple<IContinuousGenerator> exponentialGenerator
+		{
+			[Pure]
+			get => new ExponentialGeneratorDependentSimple<IContinuousGenerator>(
+				m_DependedGeneratorProvider.generator, m_Lambda);
+		}
+
+		/// <summary>
+		/// Returns a shared <see cref="ExponentialGenerator"/>.
+		/// </summary>
+		[NotNull]
+		public ExponentialGeneratorDependentSimple<IContinuousGenerator> sharedExponentialGenerator
+		{
+			get
+			{
+				if (m_sharedGenerator == null)
+				{
+					m_sharedGenerator = exponentialGenerator;
+				}
+
+				return m_sharedGenerator;
+			}
+		}
+
+		/// <summary>
+		/// Random generator that returns an independent and identically distributed random value in range [0, 1].
+		/// </summary>
+		public ContinuousGeneratorProviderReference dependedGeneratorProvider
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
+			get => m_DependedGeneratorProvider;
+			set
+			{
+				if (m_DependedGeneratorProvider == value)
+				{
+					return;
+				}
+
+				m_DependedGeneratorProvider = value;
+				m_sharedGenerator = null;
+			}
+		}
+
+		/// <summary>
+		/// Lambda. Mustn't be zero.
+		/// </summary>
+		public float lambda
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
+			get => m_Lambda;
+			set
+			{
+				if (m_Lambda == value)
+				{
+					return;
+				}
+
+				m_Lambda = value;
+				m_sharedGenerator = null;
+			}
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
+		private void OnValidate()
+		{
+			m_sharedGenerator = null;
+		}
+	}
+}